requirements.txt

# Core
pandas~=2.2.3
scikit-learn~=1.5.2

# Natural Language Processing
spacy~=3.8.2

# Data Visualization
plotly~=5.24.0
matplotlib~=3.9.2

# Langchain
langchain~=0.2.9
langchain-core~=0.2.9
langchain-community~=0.2.9

# Environment variable management
python-dotenv~=1.0.1
PyYAML~=6.0.1

numpy~=1.24.4
pathlib~=1.0.1

# Document processing
sentence-transformers~=3.0.1
